fix: use absolute path for logcat and prevent integer overflow in deleteOlderThan

+++ apps/mobile/android/app/src/main/java/com/colota/data/DatabaseHelper.kt
@@ -544,7 +544,7 @@ class DatabaseHelper private constructor(context: Context) :
     }
 
     fun deleteOlderThan(days: Int): Int {
-        val cutoff = (System.currentTimeMillis() - days * 24 * 60 * 60 * 1000L) / 1000
+        val cutoff = (System.currentTimeMillis() - days.toLong() * 24 * 60 * 60 * 1000) / 1000
         return writableDatabase.delete(TABLE_LOCATIONS, "timestamp < ?", arrayOf(cutoff.toString()))
     }
